requirements.txt
langchain-core==0.3.59
langsmith==0.3.42
langgraph==0.4.3
langchain-google-genai
python-dotenv
streamlit